+++ logger/src/main/java/org/jboss/elemento/logger/Logger.java
@@ -20,7 +20,12 @@ import java.util.Map;
 
 import elemental2.core.JsArray;
 import elemental2.core.JsDate;
+import elemental2.dom.Event;
+import elemental2.dom.Node;
+import jsinterop.annotations.JsMethod;
+import jsinterop.base.Js;
 
+import static elemental2.core.JsArray.asJsArray;
 import static elemental2.dom.DomGlobal.console;
 import static org.jboss.elemento.logger.Level.DEBUG;
 import static org.jboss.elemento.logger.Level.ERROR;
@@ -39,155 +44,256 @@ import static org.jboss.elemento.logger.Level.WARN;
  *     <li>{@link #debug(String, Object...)} → <a href="https://developer.mozilla.org/en-US/docs/Web/API/console/debug_static"><code>console.debug()</code></a></li>
  * </ol>
  * <p>
- * To get a logger use {@link Logger#getLogger(String)}. You can use an arbitrary string as category. If you use a fully qualified class name as category the package names are shortened. In any case the category is trimmed, and right aligned. <a href="https://developer.mozilla.org/en-US/docs/Web/API/console#using_string_substitutions">String substitutions</a> are supported, and you can pass arbitrary parameters to the log methods.
+ * You can use an arbitrary string as category. By using a hierarchical category, you can override subcategories. <a href="https://developer.mozilla.org/en-US/docs/Web/API/console#using_string_substitutions">String substitutions</a> are supported, and you can pass a variable list of parameters to the log methods.
  * <p>
- * The log level is set globally for all categories using {@link Logger#setLevel(Level)}.
+ * The log level is set globally for all categories using {@link Logger#setLevel(Level)}. You can override the level for one category using {@link Logger#setLevel(String, Level)}. To reset a category, use {@link Logger#resetLevel(String)}. If the category contains <code>.</code>, it is interpreted hierarchically. This means that if the category <code>org.jboss</code> is overridden, this is also applied to all subcategories (unless overridden otherwise).
  * <p>
  * The log format is predefined as
  * <pre>
- * HH:mm:ss,SSS &lt;level&gt; [&lt;category&gt;] &lt;message&gt;
+ * HH:mm:ss.SSS &lt;level&gt; [&lt;category&gt;] &lt;message&gt;
  * </pre>
+ * and cannot be customized. If the category is a fully qualified class name, the package names are shortened. In any case the category is trimmed, and right aligned.
+ * <p>
+ * The logger exports some methods with slightly adjusted signatures to JavaScript. You can use them for instance in the browser dev tools to control the global and category based log levels:
+ * <ul>
+ *     <li><code>org.jboss.elemento.logger.Logger.setLevel(String level)</code> - sets the global log level</li>
+ *     <li><code>org.jboss.elemento.logger.Logger.setLevel(String category, String level)</code> - overrides the log level for one category</li>
+ *     <li><code>org.jboss.elemento.logger.Logger.resetLevel(String category)</code> - resets the log level for the category to the global log level</li>
+ * </ul>
+ * Please use the fully qualified name!
  *
  * @see <a
  * href="https://developer.mozilla.org/en-US/docs/Web/API/console/log_static">https://developer.mozilla.org/en-US/docs/Web/API/console/log_static</a>
  */
 public class Logger {
 
+    public static final Level DEFAULT_LEVEL = INFO;
     private static final int CATEGORY_LENGTH = 23;
     private static final String ROOT_CATEGORY = "root";
+    private static final Logger ROOT_LOGGER = new Logger(ROOT_CATEGORY);
     private static final Map<String, Logger> loggers = new HashMap<>();
-    private static Level level = INFO;
+
+    // package local to enable unit tests
+    static final Map<String, Level> levels = new HashMap<>();
+    static Level level = DEFAULT_LEVEL;
+
+    static {
+        loggers.put(ROOT_CATEGORY, ROOT_LOGGER);
+    }
+
+    // ------------------------------------------------------ static API
 
     public static Logger getLogger(String name) {
-        String safeName = name == null || name.isEmpty() ? ROOT_CATEGORY : name;
-        Logger logger = loggers.get(safeName);
-        if (logger == null) {
-            logger = new Logger(safeName);
-            loggers.put(safeName, logger);
+        Logger logger;
+        if (name == null || name.isEmpty()) {
+            logger = ROOT_LOGGER;
+        } else {
+            logger = loggers.get(name);
+            if (logger == null) {
+                logger = new Logger(name);
+                loggers.put(name, logger);
+            }
         }
         return logger;
     }
 
     public static void setLevel(Level level) {
         Logger.level = level;
+        console.info("Set global log level to %s", level.name());
     }
+
+    public static void setLevel(String category, Level level) {
+        levels.put(category, level);
+        console.info("Set log level for %s to %s", category, level.name());
+    }
+
+    @JsMethod
+    public static void resetLevel(String category) {
+        levels.remove(category);
+        console.info("Log level for %s has been reset to global log level (%s)", category, Logger.level.name());
+    }
+
+    @JsMethod(name = "setLevel")
+    // String based variant for JavaScript only
+    private static void jsSetLevel(String category, String level) {
+        if (Js.typeof(level).equals("undefined")) {
+            try {
+                // just one parameter: category == level
+                setLevel(Level.valueOf(category.toUpperCase()));
+            } catch (IllegalArgumentException e) {
+                console.error("Cannot set global log level. Invalid level: '%s'", category);
+            }
+        } else {
+            try {
+                setLevel(category, Level.valueOf(level.toUpperCase()));
+            } catch (IllegalArgumentException e) {
+                console.error("Cannot set log level for '%s'. Invalid level: '%s'", category, level);
+            }
+        }
+    }
+
+    // ------------------------------------------------------ instance
 
     private final String category;
+    private final String formattedCategory;
 
     private Logger(String category) {
-        this.category = Category.format(category, CATEGORY_LENGTH);
+        this.category = category;
+        this.formattedCategory = Category.format(category, CATEGORY_LENGTH);
     }
 
+    // ------------------------------------------------------ log methods
+
     public void debug(String message, Object... params) {
-        if (level.ordinal() >= DEBUG.ordinal()) {
+        if (shouldLog(DEBUG)) {
             String formatted = format(DEBUG, message);
             if (params == null || params.length == 0) {
                 console.debug(formatted);
             } else {
-                // I wish there was another way to pass the variable arguments to console.log()
-                // But just passing 'params' does not work :-(
-                // @formatter:off
+                // I wish there was another way to pass the variable arguments to console.debug()
+                // But 'console.debug(formatted, params)' does not work as expected :-(
+                Object[] p = adjustParams(params);
                 switch (params.length) {
-                    case 1: console.debug(formatted, params[0]); break;
-                    case 2: console.debug(formatted, params[0], params[1]); break;
-                    case 3: console.debug(formatted, params[0], params[1], params[2]); break;
-                    case 4: console.debug(formatted, params[0], params[1], params[2], params[3]); break;
-                    case 5: console.debug(formatted, params[0], params[1], params[2], params[3], params[4]); break;
-                    case 6: console.debug(formatted, params[0], params[1], params[2], params[3], params[4], params[5]); break;
-                    case 7: console.debug(formatted, params[0], params[1], params[2], params[3], params[4], params[5], params[6]); break;
-                    case 8: console.debug(formatted, params[0], params[1], params[2], params[3], params[4], params[5], params[6], params[7]); break;
-                    case 9: console.debug(formatted, params[0], params[1], params[2], params[3], params[4], params[5], params[6], params[7], params[8]); break;
-                    case 10: console.debug(formatted, params[0], params[1], params[2], params[3], params[4], params[5], params[6], params[7], params[8], params[9]); break;
-                    default: console.debug(formatted, asJsArray());
+                    // @formatter:off
+                    case  1: console.debug(formatted, p[0]); break;
+                    case  2: console.debug(formatted, p[0], p[1]); break;
+                    case  3: console.debug(formatted, p[0], p[1], p[2]); break;
+                    case  4: console.debug(formatted, p[0], p[1], p[2], p[3]); break;
+                    case  5: console.debug(formatted, p[0], p[1], p[2], p[3], p[4]); break;
+                    case  6: console.debug(formatted, p[0], p[1], p[2], p[3], p[4], p[5]); break;
+                    case  7: console.debug(formatted, p[0], p[1], p[2], p[3], p[4], p[5], p[6]); break;
+                    case  8: console.debug(formatted, p[0], p[1], p[2], p[3], p[4], p[5], p[6], p[7]); break;
+                    case  9: console.debug(formatted, p[0], p[1], p[2], p[3], p[4], p[5], p[6], p[7], p[8]); break;
+                    case 10: console.debug(formatted, p[0], p[1], p[2], p[3], p[4], p[5], p[6], p[7], p[8], p[9]); break;
+                    default: console.debug(formatted, asJsArray(p));
+                    // @formatter:on
                 }
-                // @formatter:on
             }
         }
     }
 
     public void info(String message, Object... params) {
-        if (level.ordinal() >= INFO.ordinal()) {
+        if (shouldLog(INFO)) {
             String formatted = format(INFO, message);
             if (params == null || params.length == 0) {
                 console.info(formatted);
             } else {
-                // @formatter:off
+                Object[] p = adjustParams(params);
                 switch (params.length) {
-                    case 1: console.info(formatted, params[0]); break;
-                    case 2: console.info(formatted, params[0], params[1]); break;
-                    case 3: console.info(formatted, params[0], params[1], params[2]); break;
-                    case 4: console.info(formatted, params[0], params[1], params[2], params[3]); break;
-                    case 5: console.info(formatted, params[0], params[1], params[2], params[3], params[4]); break;
-                    case 6: console.info(formatted, params[0], params[1], params[2], params[3], params[4], params[5]); break;
-                    case 7: console.info(formatted, params[0], params[1], params[2], params[3], params[4], params[5], params[6]); break;
-                    case 8: console.info(formatted, params[0], params[1], params[2], params[3], params[4], params[5], params[6], params[7]); break;
-                    case 9: console.info(formatted, params[0], params[1], params[2], params[3], params[4], params[5], params[6], params[7], params[8]); break;
-                    case 10: console.info(formatted, params[0], params[1], params[2], params[3], params[4], params[5], params[6], params[7], params[8], params[9]); break;
-                    default: console.info(formatted, asJsArray());
+                    // @formatter:off
+                    case  1: console.info(formatted, p[0]); break;
+                    case  2: console.info(formatted, p[0], p[1]); break;
+                    case  3: console.info(formatted, p[0], p[1], p[2]); break;
+                    case  4: console.info(formatted, p[0], p[1], p[2], p[3]); break;
+                    case  5: console.info(formatted, p[0], p[1], p[2], p[3], p[4]); break;
+                    case  6: console.info(formatted, p[0], p[1], p[2], p[3], p[4], p[5]); break;
+                    case  7: console.info(formatted, p[0], p[1], p[2], p[3], p[4], p[5], p[6]); break;
+                    case  8: console.info(formatted, p[0], p[1], p[2], p[3], p[4], p[5], p[6], p[7]); break;
+                    case  9: console.info(formatted, p[0], p[1], p[2], p[3], p[4], p[5], p[6], p[7], p[8]); break;
+                    case 10: console.info(formatted, p[0], p[1], p[2], p[3], p[4], p[5], p[6], p[7], p[8], p[9]); break;
+                    default: console.info(formatted, asJsArray(p));
+                    // @formatter:on
                 }
-                // @formatter:on
             }
         }
     }
 
     public void warn(String message, Object... params) {
-        if (level.ordinal() >= WARN.ordinal()) {
+        if (shouldLog(WARN)) {
             String formatted = format(WARN, message);
             if (params == null || params.length == 0) {
                 console.warn(formatted);
             } else {
-                // @formatter:off
+                Object[] p = adjustParams(params);
                 switch (params.length) {
-                    case 1: console.warn(formatted, params[0]); break;
-                    case 2: console.warn(formatted, params[0], params[1]); break;
-                    case 3: console.warn(formatted, params[0], params[1], params[2]); break;
-                    case 4: console.warn(formatted, params[0], params[1], params[2], params[3]); break;
-                    case 5: console.warn(formatted, params[0], params[1], params[2], params[3], params[4]); break;
-                    case 6: console.warn(formatted, params[0], params[1], params[2], params[3], params[4], params[5]); break;
-                    case 7: console.warn(formatted, params[0], params[1], params[2], params[3], params[4], params[5], params[6]); break;
-                    case 8: console.warn(formatted, params[0], params[1], params[2], params[3], params[4], params[5], params[6], params[7]); break;
-                    case 9: console.warn(formatted, params[0], params[1], params[2], params[3], params[4], params[5], params[6], params[7], params[8]); break;
-                    case 10: console.warn(formatted, params[0], params[1], params[2], params[3], params[4], params[5], params[6], params[7], params[8], params[9]); break;
-                    default: console.warn(formatted, asJsArray());
+                    // @formatter:off
+                    case  1: console.warn(formatted, p[0]); break;
+                    case  2: console.warn(formatted, p[0], p[1]); break;
+                    case  3: console.warn(formatted, p[0], p[1], p[2]); break;
+                    case  4: console.warn(formatted, p[0], p[1], p[2], p[3]); break;
+                    case  5: console.warn(formatted, p[0], p[1], p[2], p[3], p[4]); break;
+                    case  6: console.warn(formatted, p[0], p[1], p[2], p[3], p[4], p[5]); break;
+                    case  7: console.warn(formatted, p[0], p[1], p[2], p[3], p[4], p[5], p[6]); break;
+                    case  8: console.warn(formatted, p[0], p[1], p[2], p[3], p[4], p[5], p[6], p[7]); break;
+                    case  9: console.warn(formatted, p[0], p[1], p[2], p[3], p[4], p[5], p[6], p[7], p[8]); break;
+                    case 10: console.warn(formatted, p[0], p[1], p[2], p[3], p[4], p[5], p[6], p[7], p[8], p[9]); break;
+                    default: console.warn(formatted, asJsArray(p));
+                    // @formatter:on
                 }
-                // @formatter:on
             }
         }
     }
 
     public void error(String message, Object... params) {
-        String formatted = format(ERROR, message);
-        if (params == null || params.length == 0) {
-            console.error(formatted);
-        } else {
-            // @formatter:off
-            switch (params.length) {
-                case 1: console.error(formatted, params[0]); break;
-                case 2: console.error(formatted, params[0], params[1]); break;
-                case 3: console.error(formatted, params[0], params[1], params[2]); break;
-                case 4: console.error(formatted, params[0], params[1], params[2], params[3]); break;
-                case 5: console.error(formatted, params[0], params[1], params[2], params[3], params[4]); break;
-                case 6: console.error(formatted, params[0], params[1], params[2], params[3], params[4], params[5]); break;
-                case 7: console.error(formatted, params[0], params[1], params[2], params[3], params[4], params[5], params[6]); break;
-                case 8: console.error(formatted, params[0], params[1], params[2], params[3], params[4], params[5], params[6], params[7]); break;
-                case 9: console.error(formatted, params[0], params[1], params[2], params[3], params[4], params[5], params[6], params[7], params[8]); break;
-                case 10: console.error(formatted, params[0], params[1], params[2], params[3], params[4], params[5], params[6], params[7], params[8], params[9]); break;
-                default: console.error(formatted, asJsArray());
+        if (shouldLog(ERROR)) {
+            String formatted = format(ERROR, message);
+            if (params == null || params.length == 0) {
+                console.error(formatted);
+            } else {
+                Object[] p = adjustParams(params);
+                switch (params.length) {
+                    // @formatter:off
+                    case  1: console.error(formatted, p[0]); break;
+                    case  2: console.error(formatted, p[0], p[1]); break;
+                    case  3: console.error(formatted, p[0], p[1], p[2]); break;
+                    case  4: console.error(formatted, p[0], p[1], p[2], p[3]); break;
+                    case  5: console.error(formatted, p[0], p[1], p[2], p[3], p[4]); break;
+                    case  6: console.error(formatted, p[0], p[1], p[2], p[3], p[4], p[5]); break;
+                    case  7: console.error(formatted, p[0], p[1], p[2], p[3], p[4], p[5], p[6]); break;
+                    case  8: console.error(formatted, p[0], p[1], p[2], p[3], p[4], p[5], p[6], p[7]); break;
+                    case  9: console.error(formatted, p[0], p[1], p[2], p[3], p[4], p[5], p[6], p[7], p[8]); break;
+                    case 10: console.error(formatted, p[0], p[1], p[2], p[3], p[4], p[5], p[6], p[7], p[8], p[9]); break;
+                    default: console.error(formatted, asJsArray(p));
+                    // @formatter:on
+                }
             }
-            // @formatter:on
         }
+    }
+
+    // ------------------------------------------------------ internal
+
+    boolean shouldLog(Level level) {
+        boolean global = Logger.level.ordinal() >= level.ordinal();
+        if (!Logger.levels.isEmpty()) {
+            String cc = category;
+            Level cLevel = Logger.levels.get(cc);
+            while (cLevel == null && !cc.isEmpty()) {
+                cc = Category.parent(cc);
+                cLevel = Logger.levels.get(cc);
+            }
+            if (cLevel != null) {
+                return cLevel.ordinal() >= level.ordinal();
+            }
+        }
+        return global;
     }
 
     private String format(Level level, String message) {
         String iso = new JsDate().toISOString(); // 2011-10-05T14:48:00.000Z
-        return iso.substring(11, 23) + " " + level.label + " [" + category + "] " + message;
+        return iso.substring(11, 23) + " " + level.label + " [" + formattedCategory + "] " + message;
     }
 
-    private JsArray<Object> asJsArray(Object... params) {
-        JsArray<Object> array = new JsArray<>();
-        for (Object obj : params) {
-            array.push(obj);
+    private Object[] adjustParams(Object[] params) {
+        Object[] p = new Object[params.length];
+        for (int i = 0; i < params.length; i++) {
+            if (jsNative(params[i])) {
+                p[i] = params[i];
+            } else {
+                try {
+                    p[i] = String.valueOf(params[i]);
+                } catch (Throwable t) {
+                    p[i] = "error: " + t.getMessage();
+                }
+            }
         }
-        return array;
+        return p;
+    }
+
+    private boolean jsNative(Object object) {
+        // TODO Is there a better way to detect 'native' JS objects that should be logged as is in the console?
+        return object == null || !Js.typeof(object).equals("object") ||
+                object instanceof JsArray ||
+                object instanceof Event ||
+                object instanceof Node;
     }
 }